requirements.txt
Pillow>=9.0.0
requests>=2.25.0
fastapi
uvicorn
pymongo
python-dotenv
numpy
python-multipart
flask
flask-cors 
torch 
torchvision